Added SessionDetails page with test data
Created a SessionDetails page to display the exercises and details of the chosen session.
Added a button to link from the session table row in SessionList to the details page.
Added data access layer for the Exercise model.
Added properties from the Log model to the Exercise model temporarily for testing purposes.
Registered data service in startup file.